Finished test coverage as far as we can go
git-svn-id: https://cida-svn.er.usgs.gov/repos/dev/usgs/dcpt/trunk@15628 622dc1fb-2224-0410-ab72-dc85a3975e1a

+++ src/test/java/thredds/catalog/CatalogHelperTest.java
@@ -1,9 +1,7 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
 package thredds.catalog;
 
+import java.util.List;
+import java.net.URI;
 import java.io.FileInputStream;
 import org.junit.After;
 import org.junit.Before;
@@ -14,8 +12,6 @@ import java.io.IOException;
 import java.net.URISyntaxException;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.Test;
 import static org.hamcrest.MatcherAssert.*;
 import static org.hamcrest.Matchers.*;
@@ -40,7 +36,7 @@ public class CatalogHelperTest {
 
     @After
     public void tearDownClass() throws Exception {
-        FileUtils.forceDelete(tempLocation);
+        if (tempLocation.exists()) FileUtils.forceDelete(tempLocation);
     }
     
     @Test
@@ -60,4 +56,47 @@ public class CatalogHelperTest {
         String catalog = IOUtils.toString(new FileInputStream(tempLocation));
         assertThat(catalog.contains(knownName), is(true));
     }
+    
+    /**
+     * Here we want to test that if we feed the setupCatalog() function an input of a location
+     * with an existing catalog, it won't try to set it up again
+     * 
+     * @throws URISyntaxException
+     * @throws FileNotFoundException
+     * @throws IOException 
+     */
+    @Test
+    public void testSetupCatalogUsingExistingCatalogAsInput() throws URISyntaxException, FileNotFoundException, IOException {
+        CatalogHelper.setupCatalog(tempLocation, knownName);
+        assertThat(tempLocation.exists(), is(true));
+        // We now have our catalog. But let's try to do that again with a different catalog name
+        
+        CatalogHelper.setupCatalog(tempLocation, "OH HAI I AM A CATALOG");
+        assertThat(tempLocation.exists(), is(true));
+        // At this point this catalog should not have been created
+        
+        String catalog = IOUtils.toString(new FileInputStream(tempLocation));
+        assertThat(catalog.contains(knownName), is(true));
+        assertThat(catalog.contains("OH HAI I AM A CATALOG"), is(false));
+    }
+    
+    @Test
+    public void testCreateCatalogImpl() throws URISyntaxException, FileNotFoundException, IOException {
+        URI uri = new URI("file://" + tempDir + CatalogHelper.getDefaultCatalogFilename());
+        InvCatalogImpl test = CatalogHelper.createCatalogImpl(knownName, uri);
+        assertThat(test, is(notNullValue()));
+        assertThat(test.name, is(equalTo(knownName)));
+    }
+    
+    @Test
+    public void testAddDataset() throws URISyntaxException, FileNotFoundException, IOException {
+        URI uri = new URI("file://" + tempDir + CatalogHelper.getDefaultCatalogFilename());
+        InvCatalogImpl cat = CatalogHelper.createCatalogImpl(knownName, uri);
+        InvDatasetBuilder ds = new InvDatasetBuilder("test");
+        CatalogHelper.addDataset(cat, ds.getInternalInvDataset());
+        List<InvDataset> test = cat.getDatasets();
+        assertThat(test.size(), is(equalTo(1)));
+        assertThat(test.get(0).name, is(equalTo("test")));
+    }
+    
 }
